Refactor code -- please look for comments and study the changes
