+++ frontend/src/App.js
@@ -1,0 +1,11 @@
+// eslint-disable-next-line
+import React from 'react'
+import ProfileEditor from './components/ProfileEditor'
+import Profile from './components/Profile'
+import './Styles/Styles.css'
+
+const App = () => {
+  return <Profile />
+}
+
+export default App
